🐛Refactor: AuthScreenTopView.kt file was useless, from api branch also ,removed
