--- app/src/main/java/com/zip_feast/ui/auth/screens/AuthScreenTopView.kt
@@ -1,83 +1,0 @@
-package com.zip_feast.ui.auth.screens
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import com.zip_feast.R
-import com.zip_feast.ui.theme.dimens
-
-@Composable
-fun TopSection(
-    label:String
-) {
-    val uiColor = if(isSystemInDarkTheme()) Color.White else Color.Black
-
-    Box(
-        contentAlignment = Alignment.TopCenter
-    ) {
-        Image(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(fraction = 0.46f),
-            painter = painterResource(id = R.drawable.background),
-            contentDescription = null,
-            contentScale = ContentScale.FillBounds
-        )
-
-        Row(
-            modifier = Modifier
-                .padding(top = MaterialTheme.dimens.large),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                modifier = Modifier.size(MaterialTheme.dimens.logoSize),
-                painter = painterResource(id = R.drawable.logo),
-                contentDescription = "Logo",
-                tint = uiColor
-            )
-            Spacer(modifier = Modifier.width(15.dp))
-            Column{
-                Text(
-                    text = stringResource(id = R.string.zipfeast),
-                    style = MaterialTheme.typography.headlineMedium,
-                    color = uiColor
-                )
-                Text(
-                    text = stringResource(id = R.string.zipfeast_dec),
-                    style = MaterialTheme.typography.titleMedium,
-                    color = uiColor
-                )
-
-            }
-        }
-        Text(
-            modifier = Modifier
-                .padding(bottom = 10.dp)
-                .align(alignment = Alignment.BottomCenter),
-            text = label,
-            style = MaterialTheme.typography.headlineLarge,
-            color = uiColor
-        )
-
-    }
-
-}
