+++ app/src/main/java/com/zip_feast/ui/auth/screens/LoginScreen.kt
@@ -1,5 +1,6 @@
 package com.zip_feast.ui.auth.screens
 
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -11,9 +12,12 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -25,6 +29,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -57,6 +63,64 @@ fun LoginScreen() {
             }
         }
     }
+}
+
+@Composable
+fun TopSection(
+    label: String
+) {
+
+    val uiColor = if(isSystemInDarkTheme()) Color.White else Color.Black
+
+    Box(
+        contentAlignment = Alignment.TopCenter
+    ) {
+        Image(
+            modifier = Modifier
+                .fillMaxWidth()
+                .fillMaxHeight(fraction = 0.46f),
+            painter = painterResource(id = R.drawable.background),
+            contentDescription = null,
+            contentScale = ContentScale.FillBounds
+        )
+
+        Row(
+            modifier = Modifier
+                .padding(top = MaterialTheme.dimens.large),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                modifier = Modifier.size(MaterialTheme.dimens.logoSize),
+                painter = painterResource(id = R.drawable.logo),
+                contentDescription = "Logo",
+                tint = uiColor
+            )
+            Spacer(modifier = Modifier.width(15.dp))
+            Column{
+                Text(
+                    text = stringResource(id = R.string.zipfeast),
+                    style = MaterialTheme.typography.headlineMedium,
+                    color = uiColor
+                )
+                Text(
+                    text = stringResource(id = R.string.zipfeast_dec),
+                    style = MaterialTheme.typography.titleMedium,
+                    color = uiColor
+                )
+
+            }
+        }
+        Text(
+            modifier = Modifier
+                .padding(bottom = 10.dp)
+                .align(alignment = Alignment.BottomCenter),
+            text = label,
+            style = MaterialTheme.typography.headlineLarge,
+            color = uiColor
+        )
+
+    }
+
 }
 
 @Composable
